--- test/ir/pir/cinn/test_transpose_dynshape.py
@@ -1,87 +1,0 @@
-# Copyright (c) 2025 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import unittest
-
-import numpy
-import utils
-
-import paddle
-from paddle.static import InputSpec
-
-
-class TestTransposeDynShape(unittest.TestCase):
-    def eval(self, dy_compute, inputs, input_spec):
-        dy_out = dy_compute(*inputs)
-
-        static_compute = utils.apply_to_static(
-            dy_compute, use_cinn=True, input_spec=input_spec
-        )
-        st_out = static_compute(*inputs)
-
-        for a, b in zip(
-            paddle.utils.flatten(dy_out), paddle.utils.flatten(st_out)
-        ):
-            numpy.testing.assert_allclose(a, b, atol=1e-6, rtol=1e-6)
-
-    def test_self_cross(self):
-        def func(x):
-            x_t = x.transpose([0, 2, 1])
-            return x + x_t
-
-        x = paddle.uniform([100, 200, 200])
-        x_spec = InputSpec(shape=[None, 200, 200])
-
-        self.eval(func, [x], [x_spec])
-
-    def test_small_0231(self):
-        def func(x):
-            return x.transpose([0, 2, 3, 1]) + 1
-
-        x = paddle.uniform([2, 8, 16, 160])
-        x_spec = InputSpec(shape=[2, 80, None, 160])
-
-        self.eval(func, [x], [x_spec])
-
-    def test_large_0231(self):
-        def func(x):
-            return x.transpose([0, 2, 3, 1]) + 1
-
-        x = paddle.uniform([32, 128, 14, 14])
-        x_spec = InputSpec([None, 128, None, None])
-
-        self.eval(func, [x], [x_spec])
-
-    def test_all_permuted_3210(self):
-        def func(x):
-            x = x.transpose([3, 2, 1, 0])
-            return x * (x + 1)
-
-        x = paddle.uniform([32, 14, 14, 128])
-        x_spec = InputSpec([None, None, None, 128])
-
-        self.eval(func, [x], [x_spec])
-
-    def test_unit_loop_021(self):
-        def func(x):
-            return x.transpose([0, 2, 1]) + 1
-
-        x = paddle.uniform([1024, 1, 1])
-        x_spec = InputSpec([None, None, 1])
-
-        self.eval(func, [x], [x_spec])
-
-
-if __name__ == "__main__":
-    unittest.main()
